Move the Alonzo UtxowSpec tests to a sub-module
